+++ R/rxSyntaxFunctions.R
@@ -1,6 +1,6 @@
 #' A list and description of Rode supported syntax functions
 #'
-#' @format A data frame with 3 columns and 103 rows
+#' @format A data frame with 3 columns and 106 rows
 #' \describe{
 #' \item{Function}{Reserved function Name}
 #' \item{Description}{Description of function}