+++ src/codegen.c
@@ -136,10 +136,10 @@ void prnt_vars(int scenario, int lhs, const char *pre_str, const char *post_str,
     buf = tb.ss.line[i];
     switch(scenario) {
     case print_paramLags: // Case 5 is for using #define lag_var(x)
-      printParamLags(buf, &j);
+      printParamLags(buf, &j, i);
       break;
     case print_lhsLags: // Case 4 is for using #define lag_var(x)
-      printLhsLag(buf, &j);
+      printLhsLag(buf, &j, i);
       break;
     case print_lastLhsValue: // Case 3 is for using the last lhs value
       printLastLhsValue(buf, &j);